--- src/scripts/helpers/html-element-util.ts
@@ -1,11 +1,0 @@
-export const classListContains = (element: HTMLElement, classes: string|string[]): boolean => {
-  const classArray = Array.isArray(classes) ? classes : [classes];
-
-  for (let i = 0; i < classArray.length; i++) {
-    if (element.classList.contains(classArray[i])) {
-      return true;
-    }
-  }
-
-  return false;
-};
