+++ src/scripts/staff-walkthrough-improvements/manage-walkthrough.ts
@@ -1,14 +1,10 @@
 import * as fs from 'fs';
 import { log } from 'missionlog';
+import { getDuplicates, toInt, waitForElement, allConcurrently } from '@ta-x-utilities';
+import { template, memoizeFetch } from '@ta-x-helpers';
 import { Constants } from '@ta-x-globals';
 import config from '../../config';
 import regex, { extractAllBetween } from '../../regex';
-import { waitForElement } from '../helpers/wait';
-import memoizeFetch from '../helpers/memoize-fetch';
-import { toInt } from '../helpers/parse';
-import { allConcurrently } from '../components/promise';
-import { template } from '../helpers/template';
-import { getDuplicates } from '../helpers/array-util';
 
 // Elements -------
 let walkthroughContainer: HTMLElement;
