+++ src/scripts/sticky-header.ts
@@ -1,8 +1,7 @@
 import { log } from 'missionlog';
 import { Constants } from '@ta-x-globals';
+import { toBool, waitForElement } from '@ta-x-utilities';
 import config from '../config';
-import { toBool } from './helpers/parse';
-import { waitForElement } from './helpers/wait';
 
 // Elements -------
 let extensionBody: HTMLElement;
