+++ src/utilities/array-util.ts
