+++ src/utilities/html-element-util.ts
@@ -1,3 +1,15 @@
+export const classListContains = (element: HTMLElement, classes: string|string[]): boolean => {
+  const classArray = Array.isArray(classes) ? classes : [classes];
+
+  for (let i = 0; i < classArray.length; i++) {
+    if (element.classList.contains(classArray[i])) {
+      return true;
+    }
+  }
+
+  return false;
+};
+
 export const waitForElement = (selector: string, element: Document | HTMLElement = document.documentElement, timeoutMS = 10000): Promise<HTMLElement> => new Promise(resolve => {
   if (element.querySelector(selector)) {
     return resolve(element.querySelector(selector) as HTMLElement);
@@ -25,8 +37,3 @@ export const waitForElement = (selector: string, element: Document | HTMLElement
     subtree: true
   });
 });
-
-export const until = (predFn) => {
-  const poll = (done: any) => (predFn() ? done() : setTimeout(() => poll(done), 250));
-  return new Promise(poll);
-};
